Create selenium.scraping
get_list_of_documents(idprocesso, cookies): This function takes a process ID (idprocesso) and a cookies dictionary as input. It sends an HTTP request to fetch a list of document IDs associated with the given process. The document IDs are then extracted from the response and returned as a list.

download_files(idprocesso, documentList, cookies): This function downloads the specified documents from the service. It uses the document IDs from the previous function and sends a request to initiate a multiple document download. The documents are then retrieved and stored based on the response